Split history API route-state tests
